refactor: remove flow node instance code
