+++ operate/client/src/App/ProcessInstance/index.tsx
@@ -14,7 +14,6 @@ import {useProcessInstancePageParams} from './useProcessInstancePageParams';
 import {useEffect, useRef, useState} from 'react';
 import {modificationsStore} from 'modules/stores/modifications';
 import {reaction, when} from 'mobx';
-import {flowNodeInstanceStore} from 'modules/stores/flowNodeInstance';
 import {instanceHistoryModificationStore} from 'modules/stores/instanceHistoryModification';
 import {flowNodeSelectionStore} from 'modules/stores/flowNodeSelection';
 import {flowNodeTimeStampStore} from 'modules/stores/flowNodeTimeStamp';
@@ -36,12 +35,7 @@ import {useProcessInstance} from 'modules/queries/processInstance/useProcessInst
 import {useProcessTitle} from 'modules/queries/processInstance/useProcessTitle';
 import {useCallHierarchy} from 'modules/queries/callHierarchy/useCallHierarchy';
 import {HTTP_STATUS_FORBIDDEN} from 'modules/constants/statusCode';
-import {
-  init as initFlowNodeInstance,
-  startPolling as startPollingFlowNodeInstance,
-} from 'modules/utils/flowNodeInstance';
 import {init as initFlowNodeSelection} from 'modules/utils/flowNodeSelection';
-import {type ProcessInstance as ProcessInstanceType} from '@camunda/camunda-api-zod-schemas/8.8';
 import {
   useIsRootNodeSelected,
   useRootNode,
@@ -49,14 +43,6 @@ import {
 import {notificationsStore} from 'modules/stores/notifications';
 import {useNavigate} from 'react-router-dom';
 import {Locations} from 'modules/Routes';
-
-const startPolling = (processInstance?: ProcessInstanceType) => {
-  startPollingFlowNodeInstance(processInstance, {runImmediately: true});
-};
-
-const stopPolling = () => {
-  flowNodeInstanceStore.stopPolling();
-};
 
 const ProcessInstance: React.FC = observer(() => {
   const {data: processInstance, error} = useProcessInstance();
@@ -98,28 +84,14 @@ const ProcessInstance: React.FC = observer(() => {
     const disposer = reaction(
       () => modificationsStore.isModificationModeEnabled,
       (isModificationModeEnabled) => {
-        if (isModificationModeEnabled) {
-          stopPolling();
-        } else {
+        if (!isModificationModeEnabled) {
           instanceHistoryModificationStore.reset();
-          startPolling(processInstance);
         }
       },
     );
 
-    const handleVisibilityChange = () => {
-      if (document.visibilityState === 'visible') {
-        startPolling(processInstance);
-      } else {
-        stopPolling();
-      }
-    };
-
-    document.addEventListener('visibilitychange', handleVisibilityChange);
-
     return () => {
       disposer();
-      document.removeEventListener('visibilitychange', handleVisibilityChange);
     };
   }, [processInstance]);
 
@@ -131,7 +103,6 @@ const ProcessInstance: React.FC = observer(() => {
       rootNode
     ) {
       initFlowNodeSelection(rootNode, processInstanceId, isRootNodeSelected);
-      initFlowNodeInstance(processInstance);
       isInitialized.current = true;
     }
   }, [processInstance, rootNode, processInstanceId, isRootNodeSelected]);
@@ -139,7 +110,6 @@ const ProcessInstance: React.FC = observer(() => {
   useEffect(() => {
     return () => {
       instanceHistoryModificationStore.reset();
-      flowNodeInstanceStore.reset();
       flowNodeTimeStampStore.reset();
       flowNodeSelectionStore.reset();
       modificationsStore.reset();
