+++ operate/client/src/App/ProcessInstance/tests/index.test.tsx
@@ -6,37 +6,18 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {render, screen, waitFor, within} from 'modules/testing-library';
+import {render, screen, within} from 'modules/testing-library';
 import {testData} from './index.setup';
 import {ProcessInstance} from '../index';
-import {storeStateLocally} from 'modules/utils/localStorage';
-import {flowNodeInstanceStore} from 'modules/stores/flowNodeInstance';
-import * as flowNodeInstanceUtils from 'modules/utils/flowNodeInstance';
 import {mockFetchProcessInstance as mockFetchProcessInstanceDeprecated} from 'modules/mocks/api/processInstances/fetchProcessInstance';
-import {singleInstanceMetadata} from 'modules/mocks/metadata';
-import {mockFetchFlowNodeMetadata} from 'modules/mocks/api/processInstances/fetchFlowNodeMetaData';
 import {PAGE_TITLE} from 'modules/constants';
 import {getProcessName} from 'modules/utils/instance';
-import {getWrapper, mockRequests, waitForPollingsToBeComplete} from './mocks';
+import {getWrapper, mockRequests} from './mocks';
 import {mockFetchProcessInstance} from 'modules/mocks/api/v2/processInstances/fetchProcessInstance';
 import {mockSearchVariables} from 'modules/mocks/api/v2/variables/searchVariables';
 import {createvariable} from 'modules/testUtils';
-import {act} from 'react';
 
 vi.mock('modules/utils/bpmn');
-
-const clearPollingStates = () => {
-  flowNodeInstanceStore.isPollRequestRunning = false;
-};
-
-const triggerVisibilityChange = (visibility: 'hidden' | 'visible') => {
-  Object.defineProperty(document, 'visibilityState', {
-    value: visibility,
-    configurable: true,
-  });
-
-  document.dispatchEvent(new Event('visibilitychange'));
-};
 
 describe('ProcessInstance', () => {
   beforeEach(() => {
@@ -98,57 +79,6 @@ describe('ProcessInstance', () => {
     vi.useRealTimers();
   });
 
-  it('should not trigger polling for variables when scope id changed', async () => {
-    vi.useFakeTimers({shouldAdvanceTime: true});
-
-    mockFetchFlowNodeMetadata().withSuccess(singleInstanceMetadata);
-
-    const {user} = render(<ProcessInstance />, {
-      wrapper: getWrapper({
-        selectableFlowNode: {
-          flowNodeId: 'taskD',
-          flowNodeInstanceId: 'test-id',
-        },
-      }),
-    });
-
-    storeStateLocally({
-      [`hideModificationHelperModal`]: true,
-    });
-
-    clearPollingStates();
-
-    mockRequests();
-    vi.runOnlyPendingTimers();
-
-    expect(
-      await screen.findByRole('button', {
-        name: /modify instance/i,
-      }),
-    ).toBeInTheDocument();
-
-    await user.click(
-      screen.getByRole('button', {
-        name: /modify instance/i,
-      }),
-    );
-
-    clearPollingStates();
-
-    mockRequests();
-    vi.runOnlyPendingTimers();
-
-    clearPollingStates();
-    vi.runOnlyPendingTimers();
-
-    await user.click(screen.getByRole('button', {name: 'Select flow node'}));
-
-    clearPollingStates();
-
-    vi.clearAllTimers();
-    vi.useRealTimers();
-  });
-
   it('should display forbidden content', async () => {
     mockFetchProcessInstanceDeprecated().withServerError(403);
     mockFetchProcessInstance().withServerError(403);
@@ -197,143 +127,6 @@ describe('ProcessInstance', () => {
       'href',
       'https://docs.camunda.io/docs/self-managed/operate-deployment/operate-authentication/#resource-based-permissions',
     );
-
-    vi.clearAllTimers();
-    vi.useRealTimers();
-  });
-
-  it('should stop polling if document is not visible', async () => {
-    vi.useFakeTimers({shouldAdvanceTime: true});
-
-    const initFlowNodeInstanceSpy = vi.spyOn(flowNodeInstanceUtils, 'init');
-    const startPollingFlowNodeInstanceSpy = vi.spyOn(
-      flowNodeInstanceUtils,
-      'startPolling',
-    );
-
-    mockRequests();
-    render(<ProcessInstance />, {wrapper: getWrapper()});
-
-    expect(initFlowNodeInstanceSpy).toHaveBeenCalledTimes(0);
-
-    clearPollingStates();
-    act(() => {
-      vi.runOnlyPendingTimers();
-    });
-    await waitFor(() =>
-      expect(initFlowNodeInstanceSpy).toHaveBeenCalledTimes(1),
-    );
-    act(() => {
-      vi.runOnlyPendingTimers();
-    });
-
-    await waitFor(() => {
-      expect(flowNodeInstanceStore.state.status).toBe('fetched');
-    });
-
-    triggerVisibilityChange('hidden');
-
-    clearPollingStates();
-    mockRequests();
-
-    vi.runOnlyPendingTimers();
-
-    clearPollingStates();
-    mockRequests();
-
-    vi.runOnlyPendingTimers();
-
-    expect(startPollingFlowNodeInstanceSpy).toHaveBeenCalledTimes(0);
-
-    mockRequests();
-
-    triggerVisibilityChange('visible');
-
-    clearPollingStates();
-    mockRequests();
-
-    vi.runOnlyPendingTimers();
-
-    await waitFor(() => {
-      expect(startPollingFlowNodeInstanceSpy).toHaveBeenCalledTimes(1);
-    });
-
-    await waitForPollingsToBeComplete();
-
-    vi.clearAllTimers();
-    vi.useRealTimers();
-  });
-
-  it('should not start polling in first render if document is not visible', async () => {
-    vi.useFakeTimers({shouldAdvanceTime: true});
-    triggerVisibilityChange('hidden');
-
-    const initFlowNodeInstanceSpy = vi.spyOn(flowNodeInstanceUtils, 'init');
-    const startPollingFlowNodeInstanceSpy = vi.spyOn(
-      flowNodeInstanceUtils,
-      'startPolling',
-    );
-
-    render(<ProcessInstance />, {wrapper: getWrapper()});
-
-    mockRequests();
-
-    expect(initFlowNodeInstanceSpy).toHaveBeenCalledTimes(0);
-
-    clearPollingStates();
-    await act(async () => {
-      await vi.runOnlyPendingTimersAsync();
-    });
-
-    expect(initFlowNodeInstanceSpy).toHaveBeenCalledTimes(1);
-
-    triggerVisibilityChange('visible');
-    clearPollingStates();
-
-    expect(startPollingFlowNodeInstanceSpy).toHaveBeenCalledTimes(1);
-    await act(async () => {
-      await vi.runOnlyPendingTimersAsync();
-    });
-
-    mockRequests();
-    await act(async () => {
-      await vi.runOnlyPendingTimersAsync();
-    });
-
-    expect(startPollingFlowNodeInstanceSpy).toHaveBeenCalledTimes(1);
-
-    await waitForPollingsToBeComplete();
-
-    vi.clearAllTimers();
-    vi.useRealTimers();
-  });
-
-  it('should not restart polling when document is visible for finished instances', async () => {
-    mockFetchProcessInstanceDeprecated().withSuccess(
-      testData.fetch.onPageLoad.completedProcessInstance,
-    );
-
-    vi.useFakeTimers({shouldAdvanceTime: true});
-
-    const handlePollingFlowNodeInstanceSpy = vi.spyOn(
-      flowNodeInstanceStore,
-      'pollInstances',
-    );
-
-    render(<ProcessInstance />, {wrapper: getWrapper()});
-
-    mockRequests();
-
-    expect(handlePollingFlowNodeInstanceSpy).toHaveBeenCalledTimes(0);
-
-    triggerVisibilityChange('hidden');
-    triggerVisibilityChange('visible');
-
-    clearPollingStates();
-
-    vi.runOnlyPendingTimers();
-
-    expect(handlePollingFlowNodeInstanceSpy).toHaveBeenCalledTimes(0);
 
     vi.clearAllTimers();
     vi.useRealTimers();
