+++ operate/client/src/App/ProcessInstance/tests/mocks.tsx
@@ -7,13 +7,8 @@
  */
 
 import {mockFetchProcessInstance as mockFetchProcessInstanceDeprecated} from 'modules/mocks/api/processInstances/fetchProcessInstance';
-import {mockFetchFlowNodeInstances} from 'modules/mocks/api/fetchFlowNodeInstances';
 import {testData} from './index.setup';
-import {
-  createMultiInstanceFlowNodeInstances,
-  createUser,
-  createvariable,
-} from 'modules/testUtils';
+import {createUser, createvariable} from 'modules/testUtils';
 import {createMemoryRouter, RouterProvider} from 'react-router-dom';
 import {Paths} from 'modules/Routes';
 import {LocationLog} from 'modules/utils/LocationLog';
@@ -22,9 +17,6 @@ import {
   flowNodeSelectionStore,
 } from 'modules/stores/flowNodeSelection';
 import {useEffect} from 'react';
-import {waitFor} from '@testing-library/react';
-import {processInstanceDetailsStore} from 'modules/stores/processInstanceDetails';
-import {flowNodeInstanceStore} from 'modules/stores/flowNodeInstance';
 import {mockFetchProcess} from 'modules/mocks/api/processes/fetchProcess';
 import {mockProcess} from 'modules/mocks/api/mocks/process';
 import {ProcessDefinitionKeyContext} from 'App/Processes/ListView/processDefinitionKeyContext';
@@ -42,8 +34,6 @@ import {mockMe} from 'modules/mocks/api/v2/me';
 import {mockProcessInstance} from 'modules/mocks/api/v2/mocks/processInstance';
 import {mockSearchJobs} from 'modules/mocks/api/v2/jobs/searchJobs';
 import {mockSearchIncidentsByProcessInstance} from 'modules/mocks/api/v2/incidents/searchIncidentsByProcessInstance';
-
-const processInstancesMock = createMultiInstanceFlowNodeInstances('4294980768');
 
 const mockSequenceFlowsV2: SequenceFlow[] = [
   {
@@ -96,8 +86,6 @@ const mockRequests = () => {
   mockFetchCallHierarchy().withSuccess([]);
   mockFetchProcessDefinitionXml().withSuccess('');
   mockFetchProcessSequenceFlows().withSuccess({items: mockSequenceFlowsV2});
-  mockFetchFlowNodeInstances().withSuccess(processInstancesMock.level1);
-  mockFetchFlowNodeInstances().withSuccess(processInstancesMock.level1);
   mockFetchFlownodeInstancesStatistics().withSuccess({
     items: [
       {
@@ -216,13 +204,6 @@ function getWrapper(options?: {
   return Wrapper;
 }
 
-const waitForPollingsToBeComplete = async () => {
-  await waitFor(() => {
-    expect(processInstanceDetailsStore.isPollRequestRunning).toBe(false);
-    expect(flowNodeInstanceStore.isPollRequestRunning).toBe(false);
-  });
-};
-
-export {getWrapper, waitForPollingsToBeComplete, processInstancesMock};
+export {getWrapper};
 
 export {mockRequests};
