+++ operate/client/src/modules/testUtils.tsx
@@ -6,7 +6,6 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import type {FlowNodeInstances} from 'modules/stores/flowNodeInstance';
 import type {IncidentByErrorDto} from './api/incidents/fetchIncidentsByError';
 import type {ProcessInstanceByNameDto} from './api/incidents/fetchProcessInstancesByName';
 import type {ProcessDto} from './api/processes/fetchGroupedProcesses';
@@ -1000,216 +999,6 @@ const eventSubProcess = `<?xml version="1.0" encoding="UTF-8"?>
 </bpmn:definitions>
 `;
 
-const createMultiInstanceFlowNodeInstances = (
-  processInstanceId: string,
-): {
-  level1: FlowNodeInstances;
-  level1Next: FlowNodeInstances;
-  level1Prev: FlowNodeInstances;
-  level1Poll: FlowNodeInstances;
-  level2: FlowNodeInstances;
-  level3: FlowNodeInstances;
-} => {
-  return {
-    level1: {
-      [processInstanceId]: {
-        running: null,
-        children: [
-          {
-            id: '2251799813686130',
-            type: 'PARALLEL_GATEWAY',
-            state: 'COMPLETED',
-            flowNodeId: 'peterFork',
-            startDate: '2020-08-18T12:07:33.953+0000',
-            endDate: '2020-08-18T12:07:34.034+0000',
-            treePath: `${processInstanceId}/2251799813686130`,
-            sortValues: ['1606300828415', '2251799813686130'],
-          },
-          {
-            id: '2251799813686156',
-            type: 'MULTI_INSTANCE_BODY',
-            state: 'INCIDENT',
-            flowNodeId: 'filterMapSubProcess',
-            startDate: '2020-08-18T12:07:34.205+0000',
-            endDate: null,
-            treePath: `${processInstanceId}/2251799813686156`,
-            sortValues: ['1606300828415', '2251799813686156'],
-          },
-        ],
-      },
-    },
-    level1Next: {
-      [processInstanceId]: {
-        running: null,
-        children: [
-          {
-            id: '2251799813686472',
-            type: 'PARALLEL_GATEWAY',
-            state: 'INCIDENT',
-            flowNodeId: 'filterMapSubProcess',
-            startDate: '2020-08-18T12:08:00.205+0000',
-            endDate: null,
-            treePath: `${processInstanceId}/2251799813686472`,
-            sortValues: ['1606300828415', '2251799813686472'],
-          },
-        ],
-      },
-    },
-    level1Prev: {
-      [processInstanceId]: {
-        running: null,
-        children: [
-          {
-            id: '2251390423657139',
-            type: 'START_EVENT',
-            state: 'INCIDENT',
-            flowNodeId: 'startFilterMap',
-            startDate: '2020-08-18T12:08:00.205+0000',
-            endDate: null,
-            treePath: `${processInstanceId}/2251390423657139`,
-            sortValues: ['1606300828415', '2251390423657139'],
-          },
-        ],
-      },
-    },
-    level1Poll: {
-      [processInstanceId]: {
-        running: null,
-        children: [
-          {
-            id: '2251799813686130',
-            type: 'PARALLEL_GATEWAY',
-            state: 'COMPLETED',
-            flowNodeId: 'peterFork',
-            startDate: '2020-08-18T12:07:33.953+0000',
-            endDate: '2020-08-18T12:07:34.034+0000',
-            treePath: `${processInstanceId}/2251799813686130`,
-            sortValues: ['1606300828415', '2251799813686130'],
-          },
-          {
-            id: '2251799813686156',
-            type: 'MULTI_INSTANCE_BODY',
-            state: 'COMPLETED',
-            flowNodeId: 'filterMapSubProcess',
-            startDate: '2020-08-18T12:07:34.205+0000',
-            endDate: '2020-08-18T12:07:34.034+0000',
-            treePath: `${processInstanceId}/2251799813686156`,
-            sortValues: ['1606300828415', '2251799813686156'],
-          },
-        ],
-      },
-    },
-    level2: {
-      [`${processInstanceId}/2251799813686156`]: {
-        running: true,
-        children: [
-          {
-            id: '2251799813686166',
-            type: 'SUB_PROCESS',
-            state: 'INCIDENT',
-            flowNodeId: 'filterMapSubProcess',
-            startDate: '2020-08-18T12:07:34.281+0000',
-            endDate: null,
-            treePath: `${processInstanceId}/2251799813686156/2251799813686166`,
-            sortValues: ['1606300828415', '2251799813686166'],
-          },
-        ],
-      },
-    },
-    level3: {
-      [`${processInstanceId}/2251799813686156/2251799813686166`]: {
-        running: false,
-        children: [
-          {
-            id: '2251799813686204',
-            type: 'START_EVENT',
-            state: 'COMPLETED',
-            flowNodeId: 'startFilterMap',
-            startDate: '2020-08-18T12:07:34.337+0000',
-            endDate: '2020-08-18T12:07:34.445+0000',
-            treePath: `${processInstanceId}/2251799813686156/2251799813686166/2251799813686204`,
-            sortValues: ['1606300828415', '2251799813686204'],
-          },
-        ],
-      },
-    },
-  };
-};
-
-const createEventSubProcessFlowNodeInstances = (
-  processInstanceId: string,
-): {
-  level1: FlowNodeInstances;
-  level2: FlowNodeInstances;
-} => {
-  return {
-    level1: {
-      [processInstanceId]: {
-        children: [
-          {
-            id: '6755399441057427',
-            type: 'START_EVENT',
-            state: 'COMPLETED',
-            flowNodeId: 'StartEvent_1vnazga',
-            startDate: '2021-06-22T13:43:59.698+0000',
-            endDate: '2021-06-22T13:43:59.701+0000',
-            treePath: `${processInstanceId}/6755399441057427`,
-            sortValues: ['1624369439698', '6755399441057427'],
-          },
-          {
-            id: '6755399441057429',
-            type: 'SERVICE_TASK',
-            state: 'TERMINATED',
-            flowNodeId: 'ServiceTask_1daop2o',
-            startDate: '2021-06-22T13:43:59.707+0000',
-            endDate: '2021-06-22T13:46:59.705+0000',
-            treePath: `${processInstanceId}/6755399441057429`,
-            sortValues: ['1624369439707', '6755399441057429'],
-          },
-          {
-            id: '6755399441063916',
-            type: 'EVENT_SUB_PROCESS',
-            state: 'INCIDENT',
-            flowNodeId: 'SubProcess_1ip6c6s',
-            startDate: '2021-06-22T13:46:59.705+0000',
-            endDate: null,
-            treePath: `${processInstanceId}/6755399441063916`,
-            sortValues: ['1624369619705', '6755399441063916'],
-          },
-        ],
-        running: null,
-      },
-    },
-    level2: {
-      [`${processInstanceId}/6755399441063916`]: {
-        children: [
-          {
-            id: '6755399441063918',
-            type: 'START_EVENT',
-            state: 'COMPLETED',
-            flowNodeId: 'StartEvent_1u9mwoj',
-            startDate: '2021-06-22T13:46:59.714+0000',
-            endDate: '2021-06-22T13:46:59.719+0000',
-            treePath: `${processInstanceId}/6755399441063916/6755399441063918`,
-            sortValues: ['1624369619714', '6755399441063918'],
-          },
-          {
-            id: '6755399441063920',
-            type: 'SERVICE_TASK',
-            state: 'INCIDENT',
-            flowNodeId: 'ServiceTask_0h8cwwl',
-            startDate: '2021-06-22T13:46:59.722+0000',
-            endDate: null,
-            treePath: `${processInstanceId}/6755399441063916/6755399441063920`,
-            sortValues: ['1624369619722', '6755399441063920'],
-          },
-        ],
-        running: true,
-      },
-    },
-  };
-};
-
 const mockProcessInstancesV2 = {
   items: [
     createProcessInstance({
@@ -1278,8 +1067,6 @@ export {
   eventSubProcess,
   mockProcessInstancesV2,
   mockSingleProcessInstanceV2,
-  createMultiInstanceFlowNodeInstances,
-  createEventSubProcessFlowNodeInstances,
   createvariable,
   createBatchOperation,
   createUser,
