+++ operate/client/src/modules/utils/flowNodeMetadata.ts
@@ -7,7 +7,7 @@
  */
 
 import {fetchFlowNodeMetaData} from 'modules/api/processInstances/fetchFlowNodeMetaData';
-import type {FlowNodeInstance} from 'modules/stores/flowNodeInstance';
+import type {FlowNodeInstance} from 'modules/types/operate';
 import {flowNodeMetaDataStore} from 'modules/stores/flowNodeMetaData';
 import {modificationsStore} from 'modules/stores/modifications';
 import {formatDate} from './date';
