+++ operate/client/src/modules/utils/flowNodeSelection.ts
@@ -15,10 +15,6 @@ import {
 } from 'modules/stores/flowNodeSelection';
 import {reaction, when} from 'mobx';
 import {modificationsStore} from 'modules/stores/modifications';
-import {
-  flowNodeInstanceStore,
-  type FlowNodeInstance,
-} from 'modules/stores/flowNodeInstance';
 
 /**
  * @deprecated
@@ -112,58 +108,6 @@ const selectFlowNode = (rootNode: Selection, selection: Selection) => {
  * @deprecated
  * will be migrated to useProcessInstanceElementSelection
  */
-const selectAdHocSubProcessInnerInstance = (
-  rootNode: Selection,
-  flowNodeInstance: FlowNodeInstance,
-) => {
-  const children = flowNodeInstanceStore.getVisibleChildNodes(flowNodeInstance);
-
-  const applySelection = (
-    anchorChild: FlowNodeInstance | undefined,
-    {isAnchorUpdate} = {isAnchorUpdate: false},
-  ) => {
-    if (isAnchorUpdate) {
-      const current = flowNodeSelectionStore.state.selection;
-      if (current?.flowNodeInstanceId !== flowNodeInstance.id) {
-        return; // user changed selection
-      }
-      flowNodeSelectionStore.setSelection({
-        ...current,
-        anchorFlowNodeId: anchorChild?.flowNodeId,
-      });
-      return;
-    }
-
-    selectFlowNode(rootNode, {
-      flowNodeId: flowNodeInstance.flowNodeId,
-      flowNodeInstanceId: flowNodeInstance.id,
-      flowNodeType: flowNodeInstance.type,
-      anchorFlowNodeId: anchorChild?.flowNodeId,
-    });
-  };
-
-  if (children.length > 0) {
-    applySelection(children[0]);
-  } else {
-    // initial selection without anchor
-    applySelection(undefined);
-    const disposer = when(
-      () =>
-        flowNodeInstanceStore.getVisibleChildNodes(flowNodeInstance).length > 0,
-      () => {
-        const updatedChildren =
-          flowNodeInstanceStore.getVisibleChildNodes(flowNodeInstance);
-        applySelection(updatedChildren[0], {isAnchorUpdate: true});
-        disposer();
-      },
-    );
-  }
-};
-
-/**
- * @deprecated
- * will be migrated to useProcessInstanceElementSelection
- */
 const getSelectedRunningInstanceCount = ({
   totalRunningInstancesForFlowNode,
   isRootNodeSelected,
@@ -230,7 +174,6 @@ export {
   init,
   clearSelection,
   selectFlowNode,
-  selectAdHocSubProcessInnerInstance,
   getSelectedRunningInstanceCount,
   getSelectedFlowNodeName,
 };
